Implemented password encoding and storage.
This is the meat of the program. Start with getting a master password,
needed for reading the database file entries. If you have the wrong
password, the entries get mucked, and thus unreadable.

I tried to be consistent throughout the code in preventing a password
from ever getting stored in a variable unencoded. Makes the code a big
kludgier at times, but I think that's a worth while thing.

Some code reformatting was also done, to keep the lines <80.

At this juncture, the whole storage system works. The only immediately
broken thing is the prompt interface.